+++ svc/JsonBlobService.ts
@@ -4,24 +4,52 @@
  *
  * Copyright © 2024 Extremely Heavy Industries Inc.
  */
-import {HoistService, LoadSpec, XH} from '@xh/hoist/core';
+import {HoistService, LoadSpec, PlainObject, XH} from '@xh/hoist/core';
+import {pickBy} from 'lodash';
 
 export interface JsonBlob {
-    type: string;
+    /** Either null for private blobs or special token "*" for globally shared blobs. */
+    acl: string;
+    /** True if this blob has been archived (soft-deleted). */
+    archived: boolean;
+    /** Timestamp indicating when this blob was archived, or special value `0` if not archived. */
+    archivedDate: number;
+    dateCreated: number;
+    description: string;
+    /** @internal database ID for this blob. Favor token instead. */
+    id: number;
+    lastUpdated: number;
+    lastUpdatedBy: string;
+    /** Optional application-specific metadata. */
+    meta: PlainObject | unknown[];
     name: string;
+    /** Username of the blob's creator / owner. */
+    owner: string;
+    /** Primary unique identifier for getting, updating, and archiving blobs. */
     token: string;
-    acl?: string;
-    description?: string;
+    /**
+     * Application defined type for this blob. Used as a discriminator when querying blobs related
+     * to a particular use-case within an application.
+     */
+    type: string;
+    /**
+     * Current JSON value of this blob - its contents or data. Will be undefined for blob stubs
+     * returned by `listAsync` if `includeValue` was false.
+     */
     value?: any;
-    meta?: any;
 }
 
 /**
  * Service to read and set chunks of user-specific JSON persisted via Hoist Core's JSONBlob class.
+ *
+ * This service is intended as a general, lightweight utility for persisting small bundles of
+ * unstructured data that might not warrant a full-blown domain object, but which still need to be
+ * persisted back to the database.
  */
 export class JsonBlobService extends HoistService {
     static instance: JsonBlobService;
 
+    /** Retrieve a single JSONBlob by its unique token. */
     async getAsync(token: string): Promise<JsonBlob> {
         return XH.fetchJson({
             url: 'xh/getJsonBlob',
@@ -29,16 +57,16 @@ export class JsonBlobService extends HoistService {
         });
     }
 
-    /**
-     * Return the list of blobs visible to the current user.
-     *
-     * @param type - reference key for which type of data to list.
-     * @param includeValue - true to include the full value string for each blob.
-     */
-    async listAsync(
-        {type, includeValue}: {type: string; includeValue?: boolean},
-        loadSpec: LoadSpec
-    ) {
+    /** Retrieve all blobs of a particular type that are visible to the current user. */
+    async listAsync({
+        type,
+        includeValue,
+        loadSpec
+    }: {
+        type: string;
+        includeValue?: boolean;
+        loadSpec?: LoadSpec;
+    }) {
         return XH.fetchJson({
             url: 'xh/listJsonBlobs',
             params: {type, includeValue},
@@ -48,13 +76,13 @@ export class JsonBlobService extends HoistService {
 
     /** Persist a new JSONBlob back to the server. */
     async createAsync({
-        type,
-        name,
         acl,
-        value,
+        description,
+        type,
         meta,
-        description
-    }: Omit<JsonBlob, 'token'>): Promise<JsonBlob> {
+        name,
+        value
+    }: Partial<JsonBlob>): Promise<JsonBlob> {
         return XH.fetchJson({
             url: 'xh/createJsonBlob',
             params: {
@@ -63,22 +91,23 @@ export class JsonBlobService extends HoistService {
         });
     }
 
-    /** Modify an existing JSONBlob, as identified by its unique token. */
+    /** Modify mutable properties of an existing JSONBlob, as identified by its unique token. */
     async updateAsync(
         token: string,
-        {name, acl, value, meta, description}: Omit<JsonBlob, 'token' | 'type'>
+        {acl, description, meta, name, value}: Partial<JsonBlob>
     ): Promise<JsonBlob> {
+        const update = pickBy({acl, description, meta, name, value}, (v, k) => v !== undefined);
         return XH.fetchJson({
             url: 'xh/updateJsonBlob',
             params: {
                 token,
-                update: JSON.stringify({name, acl, value, meta, description})
+                update: JSON.stringify(update)
             }
         });
     }
 
     /** Archive (soft-delete) an existing JSONBlob, as identified by its unique token. */
-    async archiveAsync(token: string) {
+    async archiveAsync(token: string): Promise<JsonBlob> {
         return XH.fetchJson({
             url: 'xh/archiveJsonBlob',
             params: {token}